suggest comment change
This proposed change is to better explain how a user identifies what their `/your_base_url` is. In our installation documentation, it is explained that this is pointed via symlink to "kora/public." In this file, But I found this comment quite confusing given my lack of prior knowledge, so this is an alternative that I hope alleviates that confusion for others.

I get this proposed comment may not be ideal either, so feel free to reject this if an alternative would work better.